% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosedir.R
\name{choose.dir}
\alias{choose.dir}
\title{Choose A Folder Interactively}
\usage{
choose.dir()
}
\description{
This function allows you to choose a folder interactively from a Finder shell folder widget on RStudio.
}
\examples{
setwd(choose.dir())
}
\keyword{choose.dir}
